# speed up pow() with exponentiating by squaring