With `-DIMAGE_WIDTH=25 -DIMAGE_HEIGHT=25`

  _  |  gcc | clang
 --- | ------ | -------
Before: | 0:03.21s - 373216K | 0:03.63s - 113836K
After: | 0:03.07s - 364412K | 0:03.48s - 112872K